requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.10=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.7.0=pypi_0
asttokens=3.0.0=pypi_0
attrs=24.3.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.11.26=h06a4308_0
certifi=2024.12.14=pypi_0
charset-normalizer=3.4.0=pypi_0
dataclasses-json=0.6.7=pypi_0
decorator=5.1.1=pypi_0
executing=2.1.0=pypi_0
frozenlist=1.5.0=pypi_0
greenlet=3.1.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.27.2=pypi_0
httpx-sse=0.4.0=pypi_0
idna=3.10=pypi_0
ipython=8.30.0=pypi_0
jedi=0.19.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
langchain=0.3.12=pypi_0
langchain-community=0.3.12=pypi_0
langchain-core=0.3.25=pypi_0
langchain-ollama=0.2.1=pypi_0
langchain-text-splitters=0.3.3=pypi_0
langgraph=0.2.59=pypi_0
langgraph-checkpoint=2.0.9=pypi_0
langgraph-sdk=0.1.46=pypi_0
langsmith=0.2.3=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
marshmallow=3.23.1=pypi_0
matplotlib-inline=0.1.7=pypi_0
msgpack=1.1.0=pypi_0
multidict=6.1.0=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
numpy=1.26.4=pypi_0
ollama=0.4.4=pypi_0
openssl=3.0.15=h5eee18b_0
orjson=3.10.12=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pillow=11.0.0=pypi_0
pip=24.2=py311h06a4308_0
prompt-toolkit=3.0.48=pypi_0
propcache=0.2.1=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pydantic=2.10.3=pypi_0
pydantic-core=2.27.1=pypi_0
pydantic-settings=2.7.0=pypi_0
pygments=2.18.0=pypi_0
python=3.11.11=he870216_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
requests-toolbelt=1.0.0=pypi_0
setuptools=75.1.0=py311h06a4308_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.36=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
tenacity=9.0.0=pypi_0
tk=8.6.14=h39e8969_0
traitlets=5.14.3=pypi_0
typing-extensions=4.12.2=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=pypi_0
wcwidth=0.2.13=pypi_0
wheel=0.44.0=py311h06a4308_0
xz=5.4.6=h5eee18b_1
yarl=1.18.3=pypi_0
zlib=1.2.13=h5eee18b_1
